AdvancedClock.
Fixed the bug causing the time factor not working.

+++ src/AdvancedClock.cpp
@@ -6,7 +6,7 @@ using namespace wp;
 void AdvancedClock::update() const
 {
     if (!m_paused)
-        m_time += static_cast<float>(m_factor)*m_chrono.restart();
+        m_time += m_factor*m_chrono.restart();
     else
         m_chrono.restart();
 }
